refactor(settings): move UI enums out of SettingsManager into Settings/Types
Separates view-shaped types (MenuBarIconStyle, HUDStyle, AppearancePreference, MenuBarPopupSize, PopupDimensions, VolumeHotkeyStep) from the persistence owner; JSON wire format unchanged. (F-ARCH-003)
